Move Shifts helper methods to Shift model
