+++ includes/view/ShiftCalendarShiftRenderer.php
@@ -3,7 +3,6 @@
 namespace Engelsystem;
 
 use Engelsystem\Config\GoodieType;
-use Engelsystem\Helpers\Shifts;
 use Engelsystem\Models\AngelType;
 use Engelsystem\Models\Shifts\Shift;
 use Engelsystem\Models\Shifts\ShiftEntry;
@@ -245,7 +244,6 @@ class ShiftCalendarShiftRenderer
      */
     private function renderShiftHead(Shift $shift, $class, $needed_angeltypes_count)
     {
-        $nightShiftsConfig = config('night_shifts');
         $goodie = GoodieType::from(config('goodie_type'));
         $goodie_enabled = $goodie !== GoodieType::None;
 
@@ -281,7 +279,7 @@ class ShiftCalendarShiftRenderer
                 ]);
         }
         $night_shift = '';
-        if (Shifts::isNightShift($shift->start, $shift->end) && $nightShiftsConfig['enabled'] && $goodie_enabled) {
+        if ($shift->isNightShift() && $goodie_enabled) {
             $night_shift = ' <i class="bi-moon-stars"></i>';
         }
