+++ includes/view/User_view.php
@@ -2,7 +2,6 @@
 
 use Carbon\Carbon;
 use Engelsystem\Config\GoodieType;
-use Engelsystem\Helpers\Shifts;
 use Engelsystem\Models\AngelType;
 use Engelsystem\Models\Group;
 use Engelsystem\Models\Shifts\Shift;
@@ -325,7 +324,7 @@ function User_view_myshift(Shift $shift, $user_source, $its_me)
     }
 
     $night_shift = '';
-    if (Shifts::isNightShift($shift->start, $shift->end) && $nightShiftsConfig['enabled'] && $goodie_enabled) {
+    if ($shift->isNightShift() && $goodie_enabled) {
         $night_shift = ' <span class="bi bi-moon-stars text-info" data-bs-toggle="tooltip" title="'
             . __('Night shifts between %d and %d am are multiplied by %d for the %s score.', [
                 $nightShiftsConfig['start'],
